@@ -10,33 +10,33 @@
   },
   "dependencies": {
     "@emotion/is-prop-valid": "latest",
-    "@radix-ui/react-accordion": "latest",
-    "@radix-ui/react-slot": "latest",
-    "@vercel/analytics": "latest",
-    "@vercel/speed-insights": "latest",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
+    "@radix-ui/react-accordion": "^1.1.2",
+    "@radix-ui/react-slot": "^1.0.2",
+    "@vercel/analytics": "^1.1.1",
+    "@vercel/speed-insights": "^1.0.12",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.0.0",
     "embla-carousel": "latest",
-    "embla-carousel-autoplay": "latest",
-    "embla-carousel-react": "latest",
-    "framer-motion": "latest",
-    "lucide-react": "^0.454.0",
-    "next": "14.0.4",
+    "embla-carousel-autoplay": "^8.0.0",
+    "embla-carousel-react": "^8.0.0",
+    "framer-motion": "^10.16.16",
+    "lucide-react": "^0.294.0",
+    "next": "14.2.35",
     "react": "^18",
     "react-dom": "^18",
-    "react-hot-toast": "latest",
-    "tailwind-merge": "^2.5.5",
+    "react-hot-toast": "^2.4.1",
+    "tailwind-merge": "^2.1.0",
     "tailwindcss-animate": "^1.0.7",
     "vue-router": "latest"
   },
   "devDependencies": {
-    "@types/node": "^22",
+    "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
     "eslint": "^8",
     "eslint-config-next": "14.0.4",
-    "postcss": "^8.5",
+    "postcss": "^8",
     "tailwindcss": "^3.3.0",
     "typescript": "^5"
   }
