@@ -2,8 +2,7 @@
 const config = {
   plugins: {
     tailwindcss: {},
-    autoprefixer: {},
   },
 }
 
-module.exports = config
+export default config
